move import functionality to TextSubmissionImportApi
